requirements.txt
django-filter  
markdown
djangorestframework
django-rest-swagger
drf-yasg